Fix imports and coverage
