+++ pyproject.toml
@@ -6,6 +6,7 @@
 
 [tool.ruff.lint]
     extend-select = ['I']
+    ignore = ['E402']
 
 [tool.ruff.lint.isort]
     combine-as-imports = true
