+++ tests/test_main.py
@@ -1,9 +1,7 @@
 import os
-from datetime import datetime
 from unittest import TestCase, mock
 
 import jwt
-import pytest
 from click.testing import CliRunner
 from fastapi import FastAPI
 from fastapi.testclient import TestClient
@@ -11,52 +9,8 @@ from requests import RequestException
 
 from tcintercom.app.logs import logfire_setup
 from tcintercom.app.main import create_app
-from tcintercom.app.routers.worker import update_duplicate_contacts
 from tcintercom.app.views import SUPPORT_TEMPLATE, conf
 from tcintercom.run import cli
-
-TEST_CONTACTS = {
-    'main_contact': {
-        'type': 'contact',
-        'id': 'main_contact',
-        'role': 'user',
-        'email': 'test_main@test.com',
-        'phone': '0123456789',
-        'created_at': datetime.timestamp(datetime(2023, 6, 1)),
-        'last_seen_at': datetime.timestamp(datetime(2023, 6, 15)),
-        'custom_attributes': {'is_duplicate': False},
-    },
-    'not_marked_duplicate_contact': {
-        'type': 'contact',
-        'id': 'duplicate_contact',
-        'role': 'user',
-        'email': 'test_main@test.com',
-        'phone': '0123456789',
-        'created_at': datetime.timestamp(datetime(2023, 6, 2)),
-        'last_seen_at': datetime.timestamp(datetime(2023, 6, 3)),
-        'custom_attributes': {'is_duplicate': False},
-    },
-    'marked_duplicate_contact': {
-        'type': 'contact',
-        'id': 'incorrect_mark_duplicate',
-        'role': 'user',
-        'email': 'test_main@test.com',
-        'phone': '0123456789',
-        'created_at': datetime.timestamp(datetime(2023, 6, 2)),
-        'last_seen_at': datetime.timestamp(datetime(2023, 6, 3)),
-        'custom_attributes': {'is_duplicate': True},
-    },
-    'created_later_contact': {
-        'type': 'contact',
-        'id': 'created_later_contact',
-        'role': 'user',
-        'email': 'test_main@test.com',
-        'phone': '0123456789',
-        'created_at': datetime.timestamp(datetime(2023, 6, 4)),
-        'last_seen_at': None,
-        'custom_attributes': {'is_duplicate': False},
-    },
-}
 
 
 def get_mock_response(test, error=False):
@@ -86,26 +40,6 @@ def get_mock_response(test, error=False):
                     return self.return_company
                 else:
                     return {'custom_attributes': {'support_plan': 'Support plan'}}
-            elif test == 'duplicate_contacts_basic':
-                if 'contacts?' in self.url:
-                    return {'data': [TEST_CONTACTS['main_contact'], TEST_CONTACTS['not_marked_duplicate_contact']]}
-            elif test == 'more_recent_not_duplicate_contact':
-                if 'contacts?' in self.url:
-                    return {'data': [TEST_CONTACTS['marked_duplicate_contact'], TEST_CONTACTS['main_contact']]}
-            elif test == 'more_recently_active_duplicate_contact':
-                if 'contacts?' in self.url:
-                    return {'data': [TEST_CONTACTS['not_marked_duplicate_contact'], TEST_CONTACTS['main_contact']]}
-            elif test == 'most_recent_created_at_duplicate_contact':
-                if self.url.endswith('/contacts?per_page=150'):
-                    return {
-                        'data': [TEST_CONTACTS['created_later_contact'], TEST_CONTACTS['main_contact']],
-                        'pages': {'next': {'starting_after': TEST_CONTACTS['main_contact'].get('id')}},
-                    }
-                elif '/contacts?per_page=150&starting_after' in self.url:
-                    return {'data': [TEST_CONTACTS['marked_duplicate_contact']]}
-            elif test == 'mark_not_duplicate_contact':
-                if 'contacts?' in self.url:
-                    return {'data': [TEST_CONTACTS['marked_duplicate_contact']]}
             elif test in self.return_dict:
                 return self.return_dict[test]
 
@@ -124,9 +58,23 @@ class TCIntercomSetup(TestCase):
     @mock.patch('tcintercom.run.uvicorn.run')
     def test_create_app(self, mock_uvicorn):
         """
-        Tests that the app is created correctly.
+        Tests that the app is created correctly when the DYNO starts with web.
         """
         os.environ['DYNO'] = 'web123'
+        runner = CliRunner()
+        result = runner.invoke(cli, 'auto')
+        assert result.exit_code == 0
+
+        assert mock_uvicorn.called
+        assert mock_uvicorn.call_count == 1
+        assert isinstance(mock_uvicorn.call_args_list[0][0][0], FastAPI)
+
+    @mock.patch('tcintercom.run.uvicorn.run')
+    def test_create_app_with_port(self, mock_uvicorn):
+        """
+        Tests that the app is created correctly when only a port is supplied.
+        """
+        os.environ['PORT'] = '8000'
         runner = CliRunner()
         result = runner.invoke(cli, 'auto')
         assert result.exit_code == 0
@@ -343,94 +291,3 @@ class BlogCallbackTestCase(TestCase):
         form_data = {'data': {'email': 'test@testing.com'}}
         r = self.client.post('/blog-callback/', json=form_data, headers={'x-webhook-signature': encoded_jwt})
         assert r.json() == {'error': 'Invalid Signature'}
-
-
-@pytest.mark.asyncio
-class TestWorker:
-    @mock.patch('tcintercom.app.views.session.request')
-    async def test_mark_duplicate_contacts(self, mock_request):
-        """
-        Tests that the correct contacts are marked as duplicate.
-        """
-        mock_request.side_effect = get_mock_response('duplicate_contacts_basic')
-        await update_duplicate_contacts({})
-
-        dup_contact = TEST_CONTACTS['not_marked_duplicate_contact']
-        assert mock_request.call_args_list[-1][0][0] == 'PUT'
-        assert mock_request.call_args_list[-1][0][1] == f'https://api.intercom.io/contacts/{dup_contact["id"]}'
-        assert mock_request.call_args_list[-1][1]['json']['email'] == dup_contact['email']
-        assert (
-            mock_request.call_args_list[-1][1]['json']['custom_attributes']['is_duplicate']
-            != dup_contact['custom_attributes']['is_duplicate']
-        )
-
-    @mock.patch('tcintercom.app.views.session.request')
-    async def test_more_recent_not_duplicate_contact(self, mock_request):
-        """
-        Tests that a more recent contact with the same email isn't incorrectly marked as duplicate due to the order
-        the contacts list is in. We assert this by checking that no calls are made to update contacts are made as they
-        are already in their correct state.
-        """
-        mock_request.side_effect = get_mock_response('more_recent_not_duplicate_contact')
-        await update_duplicate_contacts({})
-        assert mock_request.call_args_list[-1][0][0] == 'GET'
-
-    @mock.patch('tcintercom.app.views.session.request')
-    async def test_more_recently_active_contact(self, mock_request):
-        """
-        Tests that the contact who was more recently active is not marked as the duplicate contact.
-        """
-        mock_request.side_effect = get_mock_response('more_recently_active_duplicate_contact')
-        await update_duplicate_contacts({})
-
-        dup_contact = TEST_CONTACTS['not_marked_duplicate_contact']
-        assert mock_request.call_args_list[-1][0][0] == 'PUT'
-        assert mock_request.call_args_list[-1][0][1] == f'https://api.intercom.io/contacts/{dup_contact["id"]}'
-        assert mock_request.call_args_list[-1][1]['json']['email'] == dup_contact['email']
-        assert (
-            mock_request.call_args_list[-1][1]['json']['custom_attributes']['is_duplicate']
-            != dup_contact['custom_attributes']['is_duplicate']
-        )
-
-    @mock.patch('tcintercom.app.views.session.request')
-    async def test_most_recent_created_at_contact(self, mock_request):
-        """
-        Tests that a contact that a contact that was created at a later date than the original contact, gets
-        marked as the duplicate contact. Also tests the pagination of the contacts list.
-        """
-        mock_request.side_effect = get_mock_response('most_recent_created_at_duplicate_contact')
-        await update_duplicate_contacts({})
-
-        main_contact = TEST_CONTACTS['main_contact']
-        assert mock_request.call_args_list[-2][0][0] == 'GET'
-        assert (
-            mock_request.call_args_list[-2][0][1]
-            == f'https://api.intercom.io/contacts?per_page=150&starting_after={main_contact.get("id")}'
-        )
-
-        created_at_contact = TEST_CONTACTS['created_later_contact']
-        assert mock_request.call_args_list[-1][0][0] == 'PUT'
-        assert mock_request.call_args_list[-1][0][1] == f'https://api.intercom.io/contacts/{created_at_contact["id"]}'
-        assert mock_request.call_args_list[-1][1]['json']['email'] == created_at_contact['email']
-        assert (
-            mock_request.call_args_list[-1][1]['json']['custom_attributes']['is_duplicate']
-            != created_at_contact['custom_attributes']['is_duplicate']
-        )
-
-    @mock.patch('tcintercom.app.views.session.request')
-    async def test_mark_not_duplicate_contacts(self, mock_request):
-        """
-        Tests that if there is only one contact with that email address and they are marked as a duplicate,
-        we update them to be marked as not duplicate.
-        """
-        mock_request.side_effect = get_mock_response('mark_not_duplicate_contact')
-        await update_duplicate_contacts({})
-
-        dup_contact = TEST_CONTACTS['marked_duplicate_contact']
-        assert mock_request.call_args_list[-1][0][0] == 'PUT'
-        assert mock_request.call_args_list[-1][0][1] == f'https://api.intercom.io/contacts/{dup_contact["id"]}'
-        assert mock_request.call_args_list[-1][1]['json']['email'] == dup_contact['email']
-        assert (
-            mock_request.call_args_list[-1][1]['json']['custom_attributes']['is_duplicate']
-            != dup_contact['custom_attributes']['is_duplicate']
-        )
